chore: Remove Google Analytics component and add it to the RootLayout component
