+++ context/QueryModalContext.tsx
@@ -1,8 +1,6 @@
 'use client';
 // Dependencies
 import React, { createContext, useContext, useState, useEffect } from 'react';
-import { usePathname } from 'next/navigation';
-import { pageview } from '@/lib/gtag';
 
 type QueryModelContextType = {
 	isModalOpen: boolean;
@@ -22,7 +20,6 @@ const QueryModalContext = createContext({} as QueryModelContextType);
 const QueryModalProvider = ({ children }: { children: React.ReactNode }) => {
 	// State for managing the visibility of the query modal
 	const [isModalOpen, setIsModalOpen] = useState(false);
-	const pathname = usePathname();
 
 	// Opens the query modal
 	const handleOpenQueryModal = () => {
@@ -53,10 +50,6 @@ const QueryModalProvider = ({ children }: { children: React.ReactNode }) => {
 			return () => clearTimeout(timeoutId);
 		}
 	}, []);
-
-	useEffect(() => {
-		pageview(pathname);
-	}, [pathname]);
 
 	return (
 		<QueryModalContext.Provider
